+++ es74x-cloud/src/main/scala/tech/beshu/ror/proxy/es/EsActionRequestHandler.scala
@@ -1,0 +1,35 @@
+package tech.beshu.ror.proxy.es
+
+import monix.eval.Task
+import monix.execution.Scheduler
+import org.elasticsearch.action.search.SearchRequest
+import org.elasticsearch.action.{ActionRequest, ActionResponse}
+import org.elasticsearch.common.xcontent.StatusToXContentObject
+import tech.beshu.ror.proxy.es.EsActionRequestHandler.HandlingResult
+import tech.beshu.ror.proxy.es.EsActionRequestHandler.HandlingResult.{Handled, PassItThrough}
+import tech.beshu.ror.proxy.es.clients.RestHighLevelClientAdapter
+
+class EsActionRequestHandler(esClient: RestHighLevelClientAdapter)
+                            (implicit scheduler: Scheduler) {
+
+
+  def handle(request: ActionRequest): Task[HandlingResult] = {
+    tryToHandle
+      .andThen(_.map[HandlingResult](Handled))
+      .applyOrElse(request, (_: ActionRequest) => Task.now(PassItThrough))
+  }
+
+  // todo: extend it
+  private def tryToHandle: PartialFunction[ActionRequest, Task[ActionResponse with StatusToXContentObject]] = {
+    case sr: SearchRequest => esClient.search(sr)
+  }
+
+}
+
+object EsActionRequestHandler {
+  sealed trait HandlingResult
+  object HandlingResult {
+    final case class Handled(response: ActionResponse with StatusToXContentObject) extends HandlingResult
+    case object PassItThrough extends HandlingResult
+  }
+}
